Add @aws-sdk/credential-provider-node to the example
The integration example on the website does not work — nodemailer would fail to authenticate to SES unless explicitly passes a default credentials provider. Apparently, the SDK team changed the behavior in v3 to allow better tree-shaking.

I have confirmed that this configuration fixes the "Not authorized to perform ses:SendRawEmail" issue in production.